% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_function.R
\name{EM.R}
\alias{EM.R}
\title{EM Clustering}
\usage{
EM.R(x, N, it_num)
}
\arguments{
\item{x}{sample from GMM}

\item{N}{number of clusters}

\item{it_num}{number of iterations}
}
\value{
the clustering parameters and proportions
}
\description{
Implement EM algorithm to estimate the parameters and mixture proportions.
}
\examples{
\dontrun{
data_x = c(rnorm(500,0),rnorm(500,2));
M = EM.R(data_x, N = 2);M
}
}
